Move methods to command.cpp
