+++ variables.hpp
@@ -26,6 +26,26 @@ ELONA_EXTERN(elona_vector1<int> ap);
 ELONA_EXTERN(int wishfilter);
 ELONA_EXTERN(int nooracle);
 
+// command.cpp
+ELONA_EXTERN(int tile_plant);
+ELONA_EXTERN(int lv);
+ELONA_EXTERN(elona_vector1<int> moneybox);
+ELONA_EXTERN(int usermapid);
+ELONA_EXTERN(int refx);
+ELONA_EXTERN(int refy);
+ELONA_EXTERN(int enc);
+ELONA_EXTERN(int tile_trap);
+ELONA_EXTERN(int maxroute);
+ELONA_EXTERN(int movx);
+ELONA_EXTERN(int movy);
+ELONA_EXTERN(elona_vector1<int> atxid);
+ELONA_EXTERN(int prodtype);
+ELONA_EXTERN(int cellchara);
+ELONA_EXTERN(int cellfeat);
+ELONA_EXTERN(int dbg_freemove);
+ELONA_EXTERN(int ri);
+ELONA_EXTERN(int tile_downlocked);
+
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
 ELONA_EXTERN(elona_vector1<int> anix);
@@ -1013,7 +1033,7 @@ void label_1443();
 void label_1444();
 void label_1445();
 void label_1446();
-void refresh_speed_correction_value();
+void refresh_speed_correction_value(int);
 void gain_new_body_part(int);
 void gain_level(int);
 void label_1455(int);
@@ -1173,8 +1193,8 @@ void label_1886();
 void label_1888();
 void switch_religion();
 turn_result_t do_pray();
-turn_result_t do_offer();
 void label_1892();
+void start_stealing();
 void label_1901();
 void initialize_recipememory();
 void initialize_recipe();
@@ -1183,7 +1203,6 @@ void label_1933();
 void label_19342();
 void label_1935();
 turn_result_t step_into_gate();
-turn_result_t do_look_command();
 void label_1955();
 void label_1958();
 void label_1964();
@@ -1194,16 +1213,11 @@ menu_result label_1969();
 menu_result menu_journal();
 turn_result_t show_quest_board();
 void list_adventurers();
-turn_result_t do_short_cut();
-turn_result_t do_dig_command();
-turn_result_t do_bash_command();
+turn_result_t do_short_cut_command();
 turn_result_t do_use_magic();
 turn_result_t show_skill_list();
-turn_result_t do_give_command();
-turn_result_t do_interact_command();
 turn_result_t call_npc();
 void label_2022();
-turn_result_t do_cast_command();
 turn_result_t show_spell_list();
 void label_2031();
 void label_2032();
@@ -1212,20 +1226,15 @@ void label_2047(int);
 void label_2048(int);
 void equip_melee_weapon();
 menu_result menu_equipment();
-turn_result_t do_get_command();
 turn_result_t try_interact_with_npc();
 void sort_list_by_column1();
 void label_2057();
 void show_item_description();
-turn_result_t do_fire_command();
 void label_2076();
-turn_result_t do_rest_command();
 void label_2078();
-turn_result_t do_exit_command();
 void label_2081();
 turn_result_t do_gatcha();
 void dump_player_info();
-turn_result_t do_change_ammo_command();
 void remove_card_and_figures();
 void label_2088();
 void migrate_save_data();
@@ -1268,39 +1277,26 @@ void label_2161();
 void label_2162();
 void label_2187();
 void label_2188();
-turn_result_t do_throw_command();
-turn_result_t do_pray_command();
 void unequip_item(int);
 void label_2196(int);
-turn_result_t do_read_command();
-turn_result_t do_eat_command();
-turn_result_t do_drink_command();
-turn_result_t do_zap_command();
 turn_result_t do_bash();
-turn_result_t do_movement_command();
 turn_result_t proc_movement_event();
 void label_2206();
 turn_result_t do_use_stairs_command(int);
-turn_result_t do_open_command();
 void open_box();
 void open_new_year_gift();
 turn_result_t try_to_open_locked_door();
-turn_result_t do_close_command();
 void try_to_melee_attack();
 void do_physical_attack();
 void label_2220();
-turn_result_t do_search_command();
 void discover_trap();
 void discover_hidden_path();
 turn_result_t do_dig_after_sp_check();
-turn_result_t do_dip_command();
-turn_result_t do_use_command();
 turn_result_t do_plant();
 void label_2233(int);
 void label_2234(int = 0);
 void label_2235(int);
 void label_2236();
-void do_steal_command();
 void talk_start();
 void speak_to_npc();
 talk_result_t talk_more();
